production/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2019.6.16
chardet==3.0.4
Django==5.1.3
django-appconf==1.0.6
django-countries==7.6.1
django-jazzmin==2.4.9
django-redis==5.4.0
django-select2==8.2.1
djangorestframework==3.15.2
idna==2.8
Pillow==9.0.1
psycopg2==2.9.10
python-decouple==3.8
python-dotenv==1.0.1
pytz==2019.1
pyuca==1.2
redis==5.2.0
sqlparse==0.5.2
typing_extensions==4.7.1
tzdata==2024.2
whitenoise==6.8.2

gunicorn
dj-database-url
psycopg2
python-3.10.10
